Lots of changes
The code we had at the last commit did the job of scraping
content from the site fairly well, but I thought it would be
good to make it capable of dumping content straight into
another app's database.

The problem of converting websites to CMS's is pretty common,
I've seen a couple ads on craigslist for converting sites
to Wordpress. So I tried to keep robustness in mind when
I started writing these classes. For now it's geared at
Comfy Mexican Sofa, but I've been trying to write it in a
way that would make switching to other CMS's fairly simple.

A good example here is the Modeller. It connects to the database,
reads the table names, and returns the table you ask for as
a model that inherits from ActiveRecord::Base.

In turn, the Modeller doesn't know any parameters for connecting
to the database, but takes them as arguements. Most of the
database logic in the app centers around a database.yml file,
which can be generated by a rake task.

The path for the database.yml is defined as a constant in Global,
where we can also define which fields the app will recognize.
Methods throughout the app that deal with database.yml get the
field names they need from Global::DB_YML_FIELDS, and ignore all
others. Configuring database perameters for different SQL's
is just a matter of adding or removing fields from that array.

And no worries, I have been testing this with a comfy mexican sofa
app that I have running on the side. Modeller#retrieve_model('page')
will return the model CmsPage, and it's linked and ready to go.

Though dumping content into Comfy is going to be some interesting
logic. Changing CmsPage.content will change the content on the
published page, but not when you look at the page in the CMS.
The content box in the CMS view takes it's value from a CmsBlock,
with an id field linking it to that page. So for creating a new
page, we would want to create the CmsPage, and then the CmsBlock,
where we would add the content.

Interesting enough all the Cms models inherit from CmsBlock (and
ActiveRecord::Base, so no Mongo). And then there's CmsRevisions...