refactored PageFilter and wrote Configuration
PageFilter is now site-agnostic. It cares not what site it scrapes.

There is a new config file, config/nokogiri.yml. "Selectors" are
defined there, along with "element selectors" and a "seperater string".

I filled out the config file with the values appropriate for BPD.
The selector values are just Nokogiri code. The seperator string is
what the captured elements are seperated with (I have <br /> in there).

How it works:

A Configuration instance is initiated, its initiation arguement
being the nokogiri.yml file. It then assigns the field values to
instance variables. The whole instance is passed to each PageFilter
when it's initiated.

The PageFilter calls Nokogiri::HTML.parse on its html file and
assigns the Nokogiri::HTML instance to @doc.

Within a @doc.tap block, the PageFilter evaluates each selector
statement, assigning the code to @title, @meta_description and
@content. @content is then a Nokogiri::HTML instance containing
only the area of the page that we care about.

Within a @content.tap block, the PageFilter evaluates each
element selector, capturing all the elements we desire and
concatenating their .inner_html together, each element seperated
by the seperater string that we defined in nokogiri.yml.

And tada! We have a generic scraping machine.

I'm going to look into getting ActiveTest in here, so we have
a somewhat sane testing suite.