corrected lies
The whole document is a Nokogiri::HTML::Document, and any css or
xpath code on that returns a Nokogiri::XML::Element. This was
just a slight misrepresentation of the facts.

Though I'll probably refine this comment block even further to be
more accurate. The only Nokogiri object we can be certain of
is the doc that's created in PageFilter#initialize.

It makes the most sense, at least to me right now, that the content
selector should return a Nokogiri::XML::Element. But this shouldn't
be assumed either. "Element selectors" each get evaluated within
the context of the content selector, and that's the only simple
truth that should be assumed. If Joe Shmo wants the content selector
to return a String, and then have each element selector be a String
operation, there shouldn't be anything in the code prohibiting him
from doing that.

I'll modify the assertion in the PageFilter test to reflect this.