added to config file, added seperater logic
It became apparant that I had to factor down the logic even more
to get this to work.

PageFilter#parse_content no longer tries to seperate elements. It
just evaluates each element selector and appends it to a string.

This way, the user knows that the result of their element selectors
will be a string. That makes tweaking their selectors simpler.

There is another parameter in the yaml file now, elements_to_seperate.
Here the user can specify the elements they would like seperated,
like 'p' and 'img'.

This way, all captured elements aren't automatically seperated.
The purpose of the selectors is just to capture content. The
selectors are code.

elements_to_seperate is then just a list of elements to seperate.

element_selectors return one conglomerated string of html code
(the user hopes. that's up to them, not us). Then that string
is parsed for the desired elements_to_seperate, and the
seperator_string is placed between each one.

So the user can choose to capture elements a, b, c and d, but
only choose to seperate elements b and d.

I think this was a really necessary improvement.