wrote test for seperate_elements, fixed logic in yml
I discovered that it's actually to_html I want in the yml,
not inner_html. Though that was an issue with user-supplied
code, not app code. Test passes

On that note, eventually there should be some rake tasks to
help the user debug the code in their yaml. Though that is
definitely in the icebox.