added a bunch of logic for images, made corrections
I decided it was necessary to require a :root_path param for
initializing a PageFilter, in addition to the :path param.

When scraping image links on a site, they could be absolute links
or relative links. I'm not sure there's any way of handling this
logic without the PageFilter having knowledge of the root path.

I also added a line to strip out anything that matches
/http.+\.(com|org|net)/ in the image links.

One implication of this is that images hosted externally will
have their root path stripped, and return "not found" with
links indistinguishable from locally hosted images.

This happens with several images scraped from the BPD site,
because they're gifs hosted by paypal.com.

There's no logic for capturing remotely hosted images, so I think
this is totally okay. They would show up not found anyway.

The Export class doesn't raise any exceptions when it can't find
an image, but just makes a note of it and let's the user know.