strip_unwanted_characters was stripping all n's. fixed.
In the nokogiri.yml file, there's a field labeled
characters_to_strip. I had the newline character in there ("\n").

But I guess yaml relies on newlines so it was escaping it. The
value was being read into the app as "\\n". This deleted all n's.

Now the Configuration object calls map! on its characters_to_strip
array, replacing occurences of "\\n" with "\n".