linked_fields work, pages created exactly as planned
But there were things I didn't plan for, and now they're pretty
obvious.

First, when you go to edit a page in the CMS view, all of the html
is there. I haven't looked into how Comfy handles html in its
page editor. This might mean that we put the html in the page, yet
only the contents of <p> tags in the block. I dunno yet...

None of the routes were generated, so the published pages can't be
visited. This may call for a Router class.

Also processing of the linked fields may fall apart totally if there's
more than one field, or more than two models. I think this can be
solved with dynamically created IO.pipes, and a forked method to
manage access to them. In the icebox.