fixed routing by fixing the way architect handles variables
I figured out that pages weren't getting routed properly because
they had '.htm' in the slug and full_path. With '.htm' at the end,
the server looks for an actual url, the .htm file, instead of hitting
the render action on the content controller, which is what it normally
does (instead of setting a route for each page).

So I had to fix the way variables are handled in the yaml file. The user's
allowed to set variables such as '$filename' for field values in the yaml.
Previously I had a quick hack of an implementation - users could set
variables, as long as they put only the variable as the value and nothing
else. That was totally not ideal anyway, and to get the desired result
(each slug being the filename without the '.htm'), I wanted to be able to
put in the yaml something like this: $filename.sub(/\.[a-z]+$/, '').

So that works now, the variable is now substituted by the value that's
supposed to be there (so we get 'aboutus.htm'.sub(/\.[a-z]+$/, '')),
then eval'ed to produced the desired field value ("aboutus").

There are two requirements in the yaml for this to work:

1. since we're eval'ing sub on the $filename, the filename must be a string,
so we must actually write '$filename'.sub(/\.[a-z]+$/, '')

2. this makes yaml freak out about not expecting a key, so we must put
the whole thing in parentheses: ('$filename'.sub(/\.[a-z]+$/, ''))

Content is now dumped into the comfy database as desired, and the pages
are hooked up and ready to go.

Next the image saving needs fixing. I'm guessing we'll want to be able to
export all images to the 'public/images' folder in our comfy directory, then
we would want the option of automatically fixing the src's in each img tag
to chop off the directory info and only have the filename.