Removed Export class, cleaned up images logic and exporting
I started the Export class before I really developed the idea of
the Architect class, and it was never anything more than a place
to stuff "exporting logic", just for the sake of not calling FileUtils
in any of the other classes.

I decided that the Export class wasn't simplifying the overall app
design, but actually making it more complicated, and had to go. The
logic for parsing and exporting images was a mess entirely, with
methods jumping around the PageFilter, Website and Architect, all making
calls to Export. Step one in cleaning this up was getting rid of Export.

So now the images logic is pretty simple:

PageFilter#images returns an array of every img src in the page.

Website#images returns an array of every img src from all of the pages.

Architect#export_images copies all the images to an @image_folder
specified in architect.yml.

Also there is no more Website#build_site. Now building the site is
a method exclusive to the Architect class.