Refactored much of the PageFilter, page_filter.yml
Now that we had a working scraper, I wanted to push it so that
it could achieve the behavior I was looking for.

I wanted to be able to do two things:

1. Convert every p tag with the class 'lg-subhead' to an h1.

2. Operate on the src in each img tag, and call File.basename on it.

I realized that the page_filter.yml configuration could be factored down
even more, and could be even more agnostic, and I would be able to instruct
it to do these things quite easily.

So first, I got rid of elements_to_seperate and the seperator_string. They
were confusing and redundant, because we were forced to make a list of
elements in addition to the list of elements we chose to parse, and
'seperator_string' doesn't inheritly mean anything without an explanation.

Now instead, we specify actions that we want to perform on each element. So
instead of 'elements_to_seperate', we can declare 'surround_with' as an
action on the element. Using this same 'actions' interface, we can modify
attributes on the element (so we can File.basename each img src), or change
the element into another element (convert all p's with the class 'lg-subhead'
to h1's).

I also implemented the same variables concept from the architect.yml, though
for now it can only be used to reference an attribute's original value while
operating on that attribute. i.e. under "img:", we have "src: File.basename('$src')"

And I started some documentation (in './doc'), mostly to help myself think this
stuff through as I was working on it.

One thing that's broken (though all tests still pass): because we convert
the parsed content into a Nokogiri::HTML::Document to operate on it, then
call to_html on it when we're finished, it adds a verbose html tag and
some spans. Not what we want! Working on a fix.